+++ src/components/CountrySelector/__tests__/index.test.tsx
@@ -4,7 +4,7 @@ import { shallow } from 'enzyme'
 import { mockedReduxProps } from '~jest/MockedReduxProvider'
 
 import { Props as CountrySelectorProps } from '../index'
-import DocumentCountrySelector from '../DocumentCountrySelector'
+import IdentityCountrySelector from '../IdentityCountrySelector'
 
 const defaultProps: CountrySelectorProps = {
   documentType: 'driving_licence',
@@ -35,7 +35,7 @@ const defaultProps: CountrySelectorProps = {
 
 describe('CountrySelector', () => {
   it('renders without crashing', () => {
-    const wrapper = shallow(<DocumentCountrySelector {...defaultProps} />)
+    const wrapper = shallow(<IdentityCountrySelector {...defaultProps} />)
     expect(wrapper.exists()).toBeTruthy()
   })
 })