+++ src/supported-documents/index.ts
@@ -3,13 +3,14 @@ const supportedDrivingLicences = require('./supported-docs-driving_licence.json'
 const supportedNationalIDCards = require('./supported-docs-national_identity_card.json')
 const supportedResidencePermit = require('./supported-docs-residence_permit.json')
 
-import type { CountryData } from '~types/commons'
+import type { CountryData, documentSelectionType } from '~types/commons'
 import type { DocumentTypes } from '~types/steps'
 
 type SourceData = {
   country_alpha2: string
   country_alpha3: string
   country: string
+  document_type: string
 }
 
 type FlagShapes = 'rectangle' | 'square'
@@ -17,6 +18,12 @@ type FlagShapes = 'rectangle' | 'square'
 const FLAGS_FOLDER_BY_SHAPE: Record<FlagShapes, string> = {
   rectangle: '4x3',
   square: '1x1',
+}
+
+const documentTypeMap: Record<string, string> = {
+  DLD: 'driving_licence',
+  NIC: 'national_identity_card',
+  REP: 'residence_permit',
 }
 
 export const getCountryFlagSrc = (
@@ -42,6 +49,36 @@ export const getCountryDataForDocumentType = (
     return country
   }
   return null
+}
+
+export const getSupportedCountries = (
+  filterList?: documentSelectionType[]
+): CountryData[] => {
+  const allSupportedDocumentTypes = supportedDrivingLicences
+    .concat(supportedNationalIDCards)
+    .concat(supportedResidencePermit)
+  return filterList
+    ? getCountriesList(allSupportedDocumentTypes).filter((el) => {
+        return filterList.some((f) => {
+          return f.issuing_country === el.country_alpha3
+        })
+      })
+    : getCountriesList(allSupportedDocumentTypes)
+}
+
+export const getSupportedDocumentTypes = (country: string): string[] => {
+  return supportedDrivingLicences
+    .concat(supportedNationalIDCards)
+    .concat(supportedResidencePermit)
+    .filter(
+      (docData: { sourceData: SourceData }) =>
+        docData.sourceData.country_alpha3 === country
+    )
+    .map(
+      (docData: { sourceData: SourceData }) =>
+        documentTypeMap[docData.sourceData.document_type]
+    )
+    .concat(['passport'])
 }
 
 export const getSupportedCountriesForDocument = (
@@ -70,11 +107,15 @@ const getCountriesList = (supportedDocsData: { sourceData: SourceData }[]) => {
     }
   })
 
-  const uniqueCountriesList = [
-    ...new Map(
-      countriesList.map((country) => [country.country_alpha3, country])
-    ).values(),
-  ]
+  const map = new Map(
+    countriesList.map((country) => [country.country_alpha3, country])
+  )
+
+  const uniqueCountriesList = []
+
+  for (const v of map.values()) {
+    uniqueCountriesList.push(v)
+  }
 
   return uniqueCountriesList.sort((a, b) => a.name.localeCompare(b.name))
 }