+++ test/webtest/src/test/java/com/onfido/qa/websdk/test/UserConsentIT.java
@@ -1,11 +1,19 @@
 package com.onfido.qa.websdk.test;
 
 import com.onfido.qa.webdriver.common.Page;
-import com.onfido.qa.websdk.page.*;
+import com.onfido.qa.websdk.mock.Code;
+import com.onfido.qa.websdk.mock.Consent;
+import com.onfido.qa.websdk.mock.SdkConfiguration;
+import com.onfido.qa.websdk.page.CrossDeviceClientIntro;
+import com.onfido.qa.websdk.page.CrossDeviceMobileConnected;
+import com.onfido.qa.websdk.page.DocumentUpload;
+import com.onfido.qa.websdk.page.RestrictedDocumentSelection;
+import com.onfido.qa.websdk.page.UserConsent;
+import com.onfido.qa.websdk.page.Welcome;
 import org.openqa.selenium.By;
 import org.testng.annotations.Test;
 
-import java.util.Map;
+import java.util.Arrays;
 
 import static com.onfido.qa.websdk.DocumentType.PASSPORT;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -15,11 +23,17 @@ public class UserConsentIT extends WebSdkIT {
     public static final String CONSENT_FRAME_TITLE = "Onfido's privacy statement and Terms of Service";
 
     public <T extends Page> T init(Class<T> pageClass, Object... steps) {
-        return onfido().withSdkConfiguration(Map.of("sdk_features", Map.of("enable_require_applicant_consents", true))).withSteps(steps).init(pageClass);
+
+        return onfido()
+                .withSteps(steps)
+                .withMock(mock -> {
+                    mock.extend(Code.SDK_CONFIGURATION, new SdkConfiguration().withEnableRequireApplicantConsents(true));
+                })
+                .init(pageClass);
     }
 
     private UserConsent init() {
-        return init(UserConsent.class,"document");
+        return init(UserConsent.class, "document");
     }
 
     @Test(description = "should verify UI elements on the consent screen")
@@ -32,9 +46,10 @@ public class UserConsentIT extends WebSdkIT {
         verifyCopy(consent.declineButtonText(), "user_consent.button_secondary");
     }
 
+
     @Test(description = "should accept user consent")
     public void testAcceptUserConsent() {
-        var documentSelector = init().acceptUserConsent(IdDocumentSelector.class);
+        var documentSelector = init().acceptUserConsent(RestrictedDocumentSelection.class);
 
         verifyCopy(documentSelector.title(), "doc_select.title");
     }
@@ -66,7 +81,7 @@ public class UserConsentIT extends WebSdkIT {
 
     @Test(description = "remove consents from navigation once accepted")
     public void testRemoveConsentsFromNavigationOnceAccepted() {
-        var documentSelector = init().acceptUserConsent(IdDocumentSelector.class);
+        var documentSelector = init().acceptUserConsent(RestrictedDocumentSelection.class);
 
         verifyCopy(documentSelector.title(), "doc_select.title");
         assertThat(documentSelector.backArrow().isDisplayed()).isFalse();
@@ -74,41 +89,91 @@ public class UserConsentIT extends WebSdkIT {
 
     @Test(description = "remove consents from history once accepted")
     public void testRemoveConsentsFromHistoryOnceAccepted() {
-        var crossDevice = init(Welcome.class,"welcome", "document")
+        var crossDevice = init(Welcome.class, "welcome", "document")
                 .continueToNextStep(UserConsent.class)
-                .acceptUserConsent(IdDocumentSelector.class)
-                .select(PASSPORT, DocumentUpload.class)
-                     .switchToCrossDevice();
+                .acceptUserConsent(RestrictedDocumentSelection.class)
+                .selectCountry(RestrictedDocumentSelection.SUPPORTED_COUNTRY)
+                .selectDocument(PASSPORT, DocumentUpload.class)
+                .switchToCrossDevice();
 
         assertThat(crossDevice.backArrow().isDisplayed()).isTrue();
 
         var welcome = crossDevice
                 .back(DocumentUpload.class)
-                .back(IdDocumentSelector.class)
+                .back(RestrictedDocumentSelection.class)
                 .back(Welcome.class);
 
         verifyCopy(welcome.title(), "welcome.title");
     }
 
-    // @Test(description = "Is not displayed on cross device", groups = {"percy", "tabs"})
-    // public void testIsNotDisplayedOnCrossDevice() {
-    //     var link = init(Welcome.class, "welcome", "document")
-    //             .continueToNextStep(UserConsent.class)
-    //             .acceptUserConsent(RestrictedDocumentSelection.class)
-    //             .selectCountry(RestrictedDocumentSelection.SUPPORTED_COUNTRY)
-    //             .selectDocument(PASSPORT, DocumentUpload.class)
-    //             .switchToCrossDevice().getSecureLink().copyLink();
+    @Test(description = "do not show consent screen, if consent is already given")
+    public void testConsentScreenNotShownWhenConsentAlreadyGiven() {
+        var document = onfido()
+                .withSteps("document")
+                .withMock(mock -> {
+                    mock.extend(Code.SDK_CONFIGURATION, new SdkConfiguration().withEnableRequireApplicantConsents(true));
+                    mock.set(Code.CONSENTS, Arrays.asList(new Consent("privacy_notices_read_consent_given").withGranted(true)));
+                })
+                .init(RestrictedDocumentSelection.class);
 
-    //     openMobileScreen(link);
+        verifyCopy(document.title(), "doc_select.title");
+    }
 
-    //     var intro = verifyPage(CrossDeviceClientIntro.class);
-    //     switchToMainScreen();
+    @Test(description = "do not show consent screen, if consent is already given and not first screen")
+    public void testConsentScreenNotShownWhenConsentAlreadyGivenAndNotFirstScreen() {
+        var document = onfido()
+                .withSteps("welcome","document")
+                .withMock(mock -> {
+                    mock.extend(Code.SDK_CONFIGURATION, new SdkConfiguration().withEnableRequireApplicantConsents(true));
+                    mock.set(Code.CONSENTS, Arrays.asList(new Consent("privacy_notices_read_consent_given").withGranted(true)));
+                })
+                .init(Welcome.class)
+                .continueToNextStep(RestrictedDocumentSelection.class);
 
-    //     verifyPage(CrossDeviceMobileConnected.class);
-    //     switchToMobileScreen();
+        verifyCopy(document.title(), "doc_select.title");
+    }
 
-    //     var documentUpload = intro.clickContinue(DocumentUpload.class);
+    @Test(description = "Is not displayed on cross device", groups = {"percy", "tabs"})
+    public void testIsNotDisplayedOnCrossDevice() {
+        var link = init(Welcome.class, "welcome", "document")
+                .continueToNextStep(UserConsent.class)
+                .acceptUserConsent(RestrictedDocumentSelection.class)
+                .selectCountry(RestrictedDocumentSelection.SUPPORTED_COUNTRY)
+                .selectDocument(PASSPORT, DocumentUpload.class)
+                .switchToCrossDevice().getSecureLink().copyLink();
 
-    //     verifyCopy(documentUpload.title(), "doc_submit.title_passport");
-    // }
+        openMobileScreen(link);
+
+        var intro = verifyPage(CrossDeviceClientIntro.class);
+        switchToMainScreen();
+
+        verifyPage(CrossDeviceMobileConnected.class);
+        switchToMobileScreen();
+
+        var documentUpload = intro.clickContinue(DocumentUpload.class);
+
+        verifyCopy(documentUpload.title(), "doc_submit.title_passport");
+    }
+
+    @Test(description = "Is not displayed on cross device", groups = {"percy", "tabs"})
+    public void testIsNotDisplayedOnCrossDevice() {
+        var link = init(Welcome.class, "welcome", "document")
+                .continueToNextStep(UserConsent.class)
+                .acceptUserConsent(RestrictedDocumentSelection.class)
+                .selectCountry(RestrictedDocumentSelection.SUPPORTED_COUNTRY)
+                .selectDocument(PASSPORT, DocumentUpload.class)
+                .switchToCrossDevice().getSecureLink().copyLink();
+
+        openMobileScreen(link);
+
+        var intro = verifyPage(CrossDeviceClientIntro.class);
+        switchToMainScreen();
+
+        verifyPage(CrossDeviceMobileConnected.class);
+        switchToMobileScreen();
+
+        var documentUpload = intro.clickContinue(DocumentUpload.class);
+
+        verifyCopy(documentUpload.title(), "doc_submit.title_passport");
+    }
 }