+++ test/webtest/src/test/java/com/onfido/qa/websdk/test/WebSdkIT.java
@@ -7,6 +7,7 @@ import com.onfido.qa.webdriver.WebTest;
 import com.onfido.qa.webdriver.listener.BrowserStackListener;
 import com.onfido.qa.webdriver.listener.ScreenshotListener;
 import com.onfido.qa.configuration.Property;
+import com.onfido.qa.websdk.mock.Mock;
 import com.onfido.qa.websdk.sdk.WebSdk;
 import io.percy.selenium.Percy;
 import org.openqa.selenium.Point;
@@ -129,6 +130,10 @@ public abstract class WebSdkIT extends WebTest {
 
     protected WebSdk onfido() {
         return new WebSdk(driver()).withLanguage(language);
+    }
+
+    protected Mock mock() {
+        return new Mock(driver().driver);
     }
 
     protected final Copy copy(String language) {